Resume/backend/requirements.txt
fastapi
uvicorn
crewai
langchain-groq
sentence-transformers
faiss-cpu
pypdf
crewai